# api/openapi.yaml
openapi: 3.0.3
info:
  title: Monitor API
  version: "2.0.0-dev"
  description: |
    Agent and Hub HTTP API. The hand-written Flutter client in app/lib/api
    follows this contract. Authenticated routes take Authorization: Bearer.
servers:
  - url: http://127.0.0.1:19999
paths:
...
  /api/v1/operations/views:
    get:
      summary: Read the current account's personal operations views
      description: Ignores node selection. Static credentials use separate token-derived namespaces (rotation creates a new space); OIDC uses verified issuer and subject; LDAP uses directory configuration and exact login name. Anonymous and temporary share-link callers receive enabled=false and an empty collection. No owner parameter or cross-account API. Cache-Control no-store.
      responses:
        "200":
          description: revision, views, enabled, persistent, limit (10), and user (name and role only). Empty views is an array. No credentials or principal keys.
        "401":
          description: Authentication required
    post:
      summary: Replace only the current account's view collection
      description: All authenticated account roles including viewer may maintain their own filters. This does not grant alarm handling permissions. Requires the last read revision; successful writes increment it even when deleting all views. Concurrent stale requests fail without changing state. At most 10 uniquely named views per account, 1000 account spaces including deletion tombstones, and 16 MiB storage. No automatic eviction. Request body limit 32768 bytes. Cache-Control no-store.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revision, views]
              properties:
                revision: {type: integer, minimum: 0, maximum: 9007199254740991}
                views:
                  type: array
                  maxItems: 10
                  items:
                    type: object
                    additionalProperties: false
                    required: [name, severity, nodeStatus, ownerFilter, progressFilter]
                    properties:
                      name: {type: string, minLength: 1, maxLength: 40, description: Unique within this account; no leading or trailing whitespace or control characters.}
                      query: {type: string, default: "", description: At most 1024 UTF-8 bytes; no control characters.}
                      severity: {type: string, enum: [all, WARNING, CRITICAL]}
                      nodeStatus: {type: string, enum: [all, live, stale, offline]}
                      pendingOnly: {type: boolean, default: false}
                      ownerFilter: {type: string, enum: [all, mine, unassigned, assigned]}
                      progressFilter: {type: string, enum: [all, open, investigating, watching]}
      responses:
        "200":
          description: Updated collection using the GET response shape, persisted before publication.
        "400":
          description: Invalid JSON, unknown fields, missing revision, invalid filters, duplicate names or too many views
        "401":
          description: Authentication required
        "403":
          description: Anonymous or temporary share-link callers cannot persist personal views
        "409":
          description: Collection changed; reload and review before retrying
        "422":
          description: Account or file storage capacity reached
        "503":
          description: Persistence failed; in-memory collection unchanged
...
  /api/v1/operations/acknowledgements:
    post:
      summary: Acknowledge, unacknowledge or comment on the current alarm episode
      description: Admin and troubleshooter only. Actor is taken from authenticated identity. Confirmation does not silence or resolve alarms. Each episode retains the latest 20 actions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [id, action, revision]
              properties:
                id: {type: string, minLength: 64, maxLength: 64}
                action: {type: string, enum: [acknowledge, unacknowledge, comment]}
                revision: {type: integer, minimum: 0}
                note:
                  type: string
                  description: Maximum 2048 UTF-8 bytes; non-empty for comment.
      responses:
        "200":
          description: Saved record with problem context, acknowledged, revision and history.
        "400":
          description: Invalid action, note or JSON body
        "401":
          description: Authentication required
        "403":
          description: Read-only role
        "409":
          description: Episode recovered or changed, or record revision changed
        "503":
          description: Record could not be persisted
  /api/v1/operations/handling:
    post:
      summary: Assign an operator and track human investigation progress
      description: Admin and troubleshooter only. Also accepts acknowledgement actions. Actor comes from authentication. Assignment and progress do not confirm, silence or resolve the alarm. Changes apply only to the current severity episode and require its current revision. An assign target must appear in this caller's assignees list. New alarm episodes start unassigned with open progress. No external notification is sent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [id, action, revision]
              properties:
                id: {type: string, minLength: 64, maxLength: 64}
                action: {type: string, enum: [acknowledge, unacknowledge, comment, assign, unassign, progress]}
                revision: {type: integer, minimum: 0}
                note:
                  type: string
                  description: Optional except for comment; maximum 2048 UTF-8 bytes.
                assignee:
                  type: string
                  description: Required and non-empty for assign only; maximum 128 UTF-8 bytes; no control characters. Unassign clears the owner.
                status:
                  type: string
                  enum: [open, investigating, watching]
                  description: Required for progress only. Never represents alarm recovery.
      responses:
        "200":
          description: Saved record with problem, acknowledged, assignee, status, revision and recent 20 history actions. Assignment/progress actions include previous and new values.
        "400":
          description: Invalid action, fields, status, assignee or JSON
        "401":
          description: Authentication required
        "403":
          description: Read-only role
        "409":
          description: Episode recovered or changed, or record revision changed
        "503":
          description: Record could not be persisted
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security:
  - bearer: []